requirements.txt
pandas>=2.2
numpy>=1.26
pyarrow>=14.0
PyYAML>=6.0
pypdf>=5.0
